Fine tune styling for better presentation
The website looked worse than expected when viewed on my Windows
Chrome browser since I had been building it on my WSL2 Chromium
browser. I had kept this in windowed fullscreen and thus the true
presentation of the website wasn't being accurately reflected.

I tweaked the base font size and some padding / margins to allow
the website to better reflect the intended design. A link to the live
demo is now also present in the README.